--- a/tracing/build.gradle
+++ b/tracing/build.gradle
@@ -32,6 +32,9 @@ dependencies {
     implementation 'com.google.code.findbugs:jsr305'
     implementation 'com.google.errorprone:error_prone_annotations'
 
+    // TODO(dfox): this is just a placeholder
+    implementation 'io.opentelemetry:opentelemetry-exporter-logging'
+
     testImplementation "ch.qos.logback:logback-classic"
     testImplementation "com.palantir.safe-logging:preconditions-assertj"
     testImplementation "junit:junit"