--- a/tracing/src/main/java/com/palantir/tracing/DetachedSpan.java
+++ b/tracing/src/main/java/com/palantir/tracing/DetachedSpan.java
@@ -20,9 +20,6 @@ import com.google.errorprone.annotations.MustBeClosed;
 import com.palantir.logsafe.Safe;
 import com.palantir.tracing.api.SpanType;
 import io.opentelemetry.api.trace.Span;
-import io.opentelemetry.api.trace.SpanContext;
-import io.opentelemetry.api.trace.TraceFlags;
-import io.opentelemetry.api.trace.TraceState;
 import io.opentelemetry.context.Context;
 import io.opentelemetry.context.Scope;
 import java.util.Map;
@@ -78,25 +75,7 @@ public interface DetachedSpan {
             @Safe String operation,
             SpanType type) {
 
-        TraceFlags flags = TraceFlags.getDefault();
-        switch (observability) {
-            case SAMPLE:
-                flags = TraceFlags.getSampled();
-                break;
-            case DO_NOT_SAMPLE:
-                flags = TraceFlags.getDefault();
-                break;
-            case UNDECIDED:
-                // TODO(dfox): invoke our crappy sampler?
-                throw new UnsupportedOperationException();
-        }
-
-        SpanContext parentSpanContext = SpanContext.createFromRemoteParent(
-                traceId,
-                parentSpanId.orElseGet(() -> SpanContext.getInvalid().getSpanId()),
-                flags,
-                TraceState.getDefault());
-        Span parentSpan = Span.wrap(parentSpanContext);
+        Span parentSpan = Tracer.createMadeUpSpan(traceId, parentSpanId, observability);
         Context context2 = Context.current().with(parentSpan);
 
         Span span = Tracer.getSpanBuilder()