--- a/tracing/src/main/java/com/palantir/tracing/DetachedSpan.java
+++ b/tracing/src/main/java/com/palantir/tracing/DetachedSpan.java
@@ -159,7 +159,7 @@ public interface DetachedSpan {
 
         @Override
         public boolean isSampled() {
-            return span.isRecording();
+            return span.getSpanContext().isSampled();
         }
     }