… tracing-demos/src/test/java/com/palantir/tracing/BasicTracingDemos.java
… tracing-demos/src/test/java/com/palantir/tracing/TracingDemos.java
… tracing/src/main/java/com/palantir/tracing/Trace.java
… tracing/src/main/java/com/palantir/tracing/Tracers.java
… tracing/src/main/java/com/palantir/tracing/Translation.java
… tracing/src/test/java/com/palantir/tracing/CloseableTracerTest.java
… tracing/src/test/java/com/palantir/tracing/TraceTest.java
… tracing/src/test/java/com/palantir/tracing/TracerTest.java
… tracing/src/test/java/com/palantir/tracing/TracersTest.java
… versions.lock
… versions.props