## Before this PR

I had some ideas for things I wanted to try out in the tracing space (specifically around different sampling decisions), but didn't really have convenient building blocks to build them on top of. Thinking they might be easier to build on top of OpenTelemetry's abstractions instead of ours, I decided to try swapping out the underlying implementation of our library with OpenTelemetry. 

https://github.com/open-telemetry/opentelemetry-java

```




Benchmark                                                      (observability)  Mode  Cnt      Score       Error   Units
TracingBenchmark.nestedSpans                                            SAMPLE  avgt    3  40460.537 ± 14784.713   ns/op
TracingBenchmark.nestedSpans:·gc.alloc.rate                             SAMPLE  avgt    3   2692.965 ±   973.938  MB/sec
TracingBenchmark.nestedSpans:·gc.alloc.rate.norm                        SAMPLE  avgt    3  33337.922 ±     0.724    B/op
TracingBenchmark.nestedSpans:·gc.churn.G1_Eden_Space                    SAMPLE  avgt    3   2698.756 ±  1306.842  MB/sec
TracingBenchmark.nestedSpans:·gc.churn.G1_Eden_Space.norm               SAMPLE  avgt    3  33406.742 ±  4271.408    B/op
TracingBenchmark.nestedSpans:·gc.churn.G1_Survivor_Space                SAMPLE  avgt    3      0.026 ±     0.065  MB/sec
TracingBenchmark.nestedSpans:·gc.churn.G1_Survivor_Space.norm           SAMPLE  avgt    3      0.324 ±     0.709    B/op
TracingBenchmark.nestedSpans:·gc.count                                  SAMPLE  avgt    3     64.000              counts
TracingBenchmark.nestedSpans:·gc.time                                   SAMPLE  avgt    3     39.000                  ms
TracingBenchmark.nestedSpans                                     DO_NOT_SAMPLE  avgt    3   1806.962 ±    16.664   ns/op
TracingBenchmark.nestedSpans:·gc.alloc.rate                      DO_NOT_SAMPLE  avgt    3   1905.461 ±    38.852  MB/sec
TracingBenchmark.nestedSpans:·gc.alloc.rate.norm                 DO_NOT_SAMPLE  avgt    3   1000.070 ±     0.033    B/op
TracingBenchmark.nestedSpans:·gc.churn.G1_Eden_Space             DO_NOT_SAMPLE  avgt    3   1922.873 ±  1077.388  MB/sec
TracingBenchmark.nestedSpans:·gc.churn.G1_Eden_Space.norm        DO_NOT_SAMPLE  avgt    3   1009.226 ±   581.294    B/op
TracingBenchmark.nestedSpans:·gc.churn.G1_Survivor_Space         DO_NOT_SAMPLE  avgt    3      0.005 ±     0.041  MB/sec
TracingBenchmark.nestedSpans:·gc.churn.G1_Survivor_Space.norm    DO_NOT_SAMPLE  avgt    3      0.002 ±     0.021    B/op
TracingBenchmark.nestedSpans:·gc.count                           DO_NOT_SAMPLE  avgt    3     55.000              counts
TracingBenchmark.nestedSpans:·gc.time                            DO_NOT_SAMPLE  avgt    3     32.000                  ms
TracingBenchmark.nestedSpans                                         UNDECIDED  avgt    3   2535.876 ±  2356.699   ns/op
TracingBenchmark.nestedSpans:·gc.alloc.rate                          UNDECIDED  avgt    3   1786.742 ±  1598.778  MB/sec
TracingBenchmark.nestedSpans:·gc.alloc.rate.norm                     UNDECIDED  avgt    3   1347.229 ±    20.760    B/op
TracingBenchmark.nestedSpans:·gc.churn.G1_Eden_Space                 UNDECIDED  avgt    3   1784.752 ±  1882.650  MB/sec
TracingBenchmark.nestedSpans:·gc.churn.G1_Eden_Space.norm            UNDECIDED  avgt    3   1345.367 ±   306.507    B/op
TracingBenchmark.nestedSpans:·gc.churn.G1_Survivor_Space             UNDECIDED  avgt    3      0.016 ±     0.081  MB/sec
TracingBenchmark.nestedSpans:·gc.churn.G1_Survivor_Space.norm        UNDECIDED  avgt    3      0.012 ±     0.062    B/op
TracingBenchmark.nestedSpans:·gc.count                               UNDECIDED  avgt    3     51.000              counts
TracingBenchmark.nestedSpans:·gc.time                                UNDECIDED  avgt    3     30.000                  ms

```

## After this PR
==COMMIT_MSG==
OpenTelemetry
==COMMIT_MSG==

```
45x slower
1218x slower
625x slower

Benchmark                                                      (observability)  Mode  Cnt        Score         Error   Units
TracingBenchmark.nestedSpans                                            SAMPLE  avgt    3  1809725.839 ± 1116885.217   ns/op
TracingBenchmark.nestedSpans:·gc.alloc.rate                             SAMPLE  avgt    3      539.890 ±     408.617  MB/sec
TracingBenchmark.nestedSpans:·gc.alloc.rate.norm                        SAMPLE  avgt    3   296089.478 ±   21532.665    B/op
TracingBenchmark.nestedSpans:·gc.churn.G1_Eden_Space                    SAMPLE  avgt    3      541.898 ±     906.463  MB/sec
TracingBenchmark.nestedSpans:·gc.churn.G1_Eden_Space.norm               SAMPLE  avgt    3   297834.870 ±  625864.610    B/op
TracingBenchmark.nestedSpans:·gc.churn.G1_Survivor_Space                SAMPLE  avgt    3        0.030 ±       0.413  MB/sec
TracingBenchmark.nestedSpans:·gc.churn.G1_Survivor_Space.norm           SAMPLE  avgt    3       16.013 ±     212.809    B/op
TracingBenchmark.nestedSpans:·gc.count                                  SAMPLE  avgt    3       19.000                counts
TracingBenchmark.nestedSpans:·gc.time                                   SAMPLE  avgt    3       10.000                    ms
TracingBenchmark.nestedSpans                                     DO_NOT_SAMPLE  avgt    3  2201364.233 ± 7974683.261   ns/op
TracingBenchmark.nestedSpans:·gc.alloc.rate                      DO_NOT_SAMPLE  avgt    3      428.575 ±    1572.230  MB/sec
TracingBenchmark.nestedSpans:·gc.alloc.rate.norm                 DO_NOT_SAMPLE  avgt    3   262455.205 ±   21528.414    B/op
TracingBenchmark.nestedSpans:·gc.churn.G1_Eden_Space             DO_NOT_SAMPLE  avgt    3      427.792 ±    1562.025  MB/sec
TracingBenchmark.nestedSpans:·gc.churn.G1_Eden_Space.norm        DO_NOT_SAMPLE  avgt    3   262109.386 ±   81942.723    B/op
TracingBenchmark.nestedSpans:·gc.churn.G1_Survivor_Space         DO_NOT_SAMPLE  avgt    3        0.012 ±       0.213  MB/sec
TracingBenchmark.nestedSpans:·gc.churn.G1_Survivor_Space.norm    DO_NOT_SAMPLE  avgt    3        7.404 ±     126.659    B/op
TracingBenchmark.nestedSpans:·gc.count                           DO_NOT_SAMPLE  avgt    3       15.000                counts
TracingBenchmark.nestedSpans:·gc.time                            DO_NOT_SAMPLE  avgt    3        8.000                    ms
TracingBenchmark.nestedSpans                                         UNDECIDED  avgt    3  1585828.421 ±   67290.043   ns/op
TracingBenchmark.nestedSpans:·gc.alloc.rate                          UNDECIDED  avgt    3      560.346 ±      49.638  MB/sec
TracingBenchmark.nestedSpans:·gc.alloc.rate.norm                     UNDECIDED  avgt    3   266744.604 ±   21632.544    B/op
TracingBenchmark.nestedSpans:·gc.churn.G1_Eden_Space                 UNDECIDED  avgt    3      570.355 ±     896.129  MB/sec
TracingBenchmark.nestedSpans:·gc.churn.G1_Eden_Space.norm            UNDECIDED  avgt    3   271501.465 ±  424465.246    B/op
TracingBenchmark.nestedSpans:·gc.churn.G1_Survivor_Space             UNDECIDED  avgt    3        0.018 ±       0.158  MB/sec
TracingBenchmark.nestedSpans:·gc.churn.G1_Survivor_Space.norm        UNDECIDED  avgt    3        8.769 ±      75.439    B/op
TracingBenchmark.nestedSpans:·gc.count                               UNDECIDED  avgt    3       20.000                counts
TracingBenchmark.nestedSpans:·gc.time                                UNDECIDED  avgt    3       10.000                    ms
```


## Possible downsides?
<!-- Please describe any way users could be negatively affected by this PR. -->